requirements.txt
playwright
behave
allure-behave
allure-python-commons